+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Removed unnecessary documentation. I will be relying on the README files.
 - Tweaked the plugins to be a bit more robust in display of detailed descriptions.
 - Specifying "login_user" in the people plugin now also triggers "show_details".
+- Added the "serverinfo" response (God login only).
 
 **1.0.0.2000** *(July 6, 2018)*